@@ -1,6 +1,6 @@
 package com.lazahata.myhp.ui.common;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 /**
  * Created by lazahata on 13/2/2017.
