@@ -1,6 +1,6 @@
 package com.lazahata.core.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
